+++ Example/VLCRemoteKitExample/ViewController.swift
@@ -1,0 +1,23 @@
+//
+//  ViewController.swift
+//  VLCRemoteKitExample
+//
+//  Created by Yannick LORIOT on 17/07/15.
+//  Copyright (c) 2015 Yannick LORIOT. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    // Do any additional setup after loading the view, typically from a nib.
+  }
+
+  override func didReceiveMemoryWarning() {
+    super.didReceiveMemoryWarning()
+    // Dispose of any resources that can be recreated.
+  }
+}
+