+++ VLCRemoteKit/VLCRemoteKit.swift
@@ -1,0 +1,9 @@
+//
+//  VLCRemoteKit.swift
+//  VLCRemoteKitExample
+//
+//  Created by Yannick LORIOT on 17/07/15.
+//  Copyright (c) 2015 Yannick LORIOT. All rights reserved.
+//
+
+import Foundation